+++ lib/fields/helpers/combo_page.dart
@@ -1,9 +1,9 @@
 import 'package:every_door/constants.dart';
 import 'package:every_door/fields/combo.dart';
+import 'package:every_door/helpers/normalizer.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_grid_list/responsive_grid_list.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import "package:unorm_dart/unorm_dart.dart" as unorm;
 
 class ComboChooserPage extends StatefulWidget {
   final ComboPresetField field;
@@ -83,19 +83,14 @@ class _ComboChooserPageState extends State<ComboChooserPage> {
     );
   }
 
-  String _normalize(String s) {
-    var combining = RegExp(r"[\u0300-\u036F]");
-    return unorm.nfkd(s.toLowerCase().trim()).replaceAll(combining, '');
-  }
-
   Widget buildChooser(BuildContext context) {
     final loc = AppLocalizations.of(context)!;
     var options = missingOptions + List.of(widget.field.options);
     if (filter.isNotEmpty) {
       // Prune options
-      final nFilter = _normalize(filter);
+      final nFilter = normalizeString(filter);
       options = options
-          .where((element) => _normalize(element.value).contains(nFilter))
+          .where((element) => normalizeString(element.value).contains(nFilter))
           .toList();
       if (widget.field.customValues) {
         // Add this new value as an option